# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /feed:
    get:
      operationId: listPosts
      tags: [feed]
      summary: List community posts
      parameters:
        - name: communityId
          in: query
          required: false
          schema:
            type: string
            default: "default"
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostList"
    post:
      operationId: createPost
      tags: [feed]
      summary: Create a community post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostInput"
      responses:
        "201":
          description: Created post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
...
  /admin/members:
    get:
      operationId: adminListMembers
      tags: [admin]
      summary: List community members
      parameters:
        - name: communityId
          in: query
          required: false
          schema:
            type: string
            default: "default"
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Member"
    post:
      operationId: adminCreateMember
      tags: [admin]
      summary: Submit a join request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberInput"
      responses:
        "201":
          description: Created member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
  /admin/members/{memberId}/approve:
    patch:
      operationId: adminApproveMember
      tags: [admin]
      parameters:
        - name: memberId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
  /admin/members/{memberId}/reject:
    patch:
      operationId: adminRejectMember
      tags: [admin]
      parameters:
        - name: memberId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
  /admin/members/{memberId}/verify:
    patch:
      operationId: adminVerifyMember
      tags: [admin]
      parameters:
        - name: memberId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
  /admin/members/{memberId}/role:
    patch:
      operationId: adminSetMemberRole
      tags: [admin]
      parameters:
        - name: memberId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberRoleUpdate"
      responses:
        "200":
          description: Updated member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
  /admin/members/{memberId}:
    delete:
      operationId: adminDeleteMember
      tags: [admin]
      parameters:
        - name: memberId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
  /admin/posts:
    get:
      operationId: adminListPosts
      tags: [admin]
      summary: List all posts for admin
      parameters:
        - name: communityId
          in: query
          required: false
          schema:
            type: string
            default: "default"
      responses:
        "200":
          description: List of posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
  /admin/posts/{postId}:
    delete:
      operationId: adminDeletePost
      tags: [admin]
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
  /admin/posts/{postId}/pin:
    patch:
      operationId: adminTogglePin
      tags: [admin]
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /admin/community:
    get:
      operationId: adminGetCommunity
      tags: [admin]
      parameters:
        - name: communityId
          in: query
          required: false
          schema:
            type: string
            default: "default"
      responses:
        "200":
          description: Community settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommunitySettings"
    put:
      operationId: adminUpdateCommunity
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommunitySettingsInput"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommunitySettings"
  /admin/stats:
    get:
      operationId: adminGetStats
      tags: [admin]
      parameters:
        - name: communityId
          in: query
          required: false
          schema:
            type: string
            default: "default"
      responses:
        "200":
          description: Community stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
  /admin/announcements:
    post:
      operationId: adminCreateAnnouncement
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnouncementInput"
      responses:
        "201":
          description: Created announcement post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
...
components:
  schemas:
...
    Post:
      type: object
      required:
        - id
        - communityId
        - userId
        - userName
        - unitNumber
        - type
        - title
        - body
        - imageUrls
        - isPinned
        - likesCount
        - commentsCount
        - createdAt
      properties:
        id:
          type: integer
        communityId:
          type: string
        userId:
          type: string
        userName:
          type: string
        unitNumber:
          type: string
        type:
          type: string
        title:
          type: string
        body:
          type: string
        imageUrls:
          type: array
          items:
            type: string
        isPinned:
          type: boolean
        likesCount:
          type: integer
        commentsCount:
          type: integer
        createdAt:
          type: string
    PostList:
      type: object
      required:
        - posts
        - total
        - page
        - limit
        - hasMore
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        hasMore:
          type: boolean
    PostInput:
      type: object
      required:
        - title
        - body
        - type
      properties:
        communityId:
          type: string
          default: "default"
        userId:
          type: string
          default: "ahmed"
        userName:
          type: string
          default: "Ahmed Khan"
        unitNumber:
          type: string
          default: "B-204"
        type:
          type: string
        title:
          type: string
        body:
          type: string
        imageUrls:
          type: array
          items:
            type: string
        isPinned:
          type: boolean
          default: false
...
    Member:
      type: object
      required: [id, communityId, userId, name, unitNumber, phone, status, role, isVerified, joinDate]
      properties:
        id:
          type: integer
        communityId:
          type: string
        userId:
          type: string
        name:
          type: string
        unitNumber:
          type: string
        phone:
          type: string
        status:
          type: string
        role:
          type: string
        isVerified:
          type: boolean
        joinDate:
          type: string
    MemberRoleUpdate:
      type: object
      required: [role]
      properties:
        role:
          type: string
    MemberInput:
      type: object
      required: [userId, name, unitNumber]
      properties:
        communityId:
          type: string
          default: "default"
        userId:
          type: string
        name:
          type: string
        unitNumber:
          type: string
        phone:
          type: string
          default: ""
        role:
          type: string
          default: "resident"
    CommunitySettings:
      type: object
      required: [id, communityId, name, area, city, rules]
      properties:
        id:
          type: integer
        communityId:
          type: string
        name:
          type: string
        area:
          type: string
        city:
          type: string
        logoUrl:
          type: string
          nullable: true
        rules:
          type: string
    CommunitySettingsInput:
      type: object
      properties:
        communityId:
          type: string
          default: "default"
        name:
          type: string
        area:
          type: string
        city:
          type: string
        logoUrl:
          type: string
          nullable: true
        rules:
          type: string
    AdminStats:
      type: object
      required: [totalMembers, postsThisMonth, activeListings, pendingMembers]
      properties:
        totalMembers:
          type: integer
        postsThisMonth:
          type: integer
        activeListings:
          type: integer
        pendingMembers:
          type: integer
    AnnouncementInput:
      type: object
      required: [title, body]
      properties:
        communityId:
          type: string
          default: "default"
        title:
          type: string
        body:
          type: string
        userId:
          type: string
          default: "ahmed"
        userName:
          type: string
          default: "Ahmed Khan"
        unitNumber:
          type: string
          default: "B-204"
...